move: component files into component folder
