+++ app/[locale]/[[...path]]/page.tsx
@@ -1,4 +1,4 @@
-import { Header } from "@/stories/header/Header";
+import { Header } from "components/molecules/header/Header";
 
 export default function Home() {
   return (
